+++ monty.h
@@ -1,6 +1,5 @@
 #ifndef MONTY_H
 #define MONTY_H
-
 #include <stdio.h>
 #include <stdlib.h>
 #include <sys/types.h>
@@ -8,48 +7,29 @@
 #include <fcntl.h>
 #include <string.h>
 #include <ctype.h>
-
 /**
-* struct stack_s - doubly linked list representation of a stack (or queue)
-* @n: integer
-* @prev: points to the previous element of the stack (or queue)
-* @next: points to the next element of the stack (or queue)
-*
-* Description: doubly linked list node structure
-* for stack, queues, LIFO, FIFO
-*/
+ * struct stack_s - doubly linked list representation of a stack (or queue)
+ * @n: integer
+ * @prev: points to the previous element of the stack (or queue)
+ * @next: points to the next element of the stack (or queue)
+ *
+ * Description: doubly linked list node structure
+ * for stack, queues, LIFO, FIFO Holberton project
+ */
 typedef struct stack_s
 {
 	int n;
 	struct stack_s *prev;
 	struct stack_s *next;
 } stack_t;
-
-
 /**
-* struct instruction_s - opcode and its function
-* @opcode: the opcode
-* @f: function to handle the opcode
-*
-* Description: opcode and its function
-* for stack, queues, LIFO, FIFO
-*/
-typedef struct instruction_s
-{
-	char *opcode;
-	void (*f)(stack_t **stack, unsigned int line_number);
-} instruction_t;
-
-
-/**
-* struct bus_s - variables -args, file, line content
-* @arg: value
-* @file: pointer to monty file
-* @content: line content
-* @lifi: flag change stack <-> queue
-*
-* Description: carries values through the program
-*/
+ * struct bus_s - variables -args, file, line content
+ * @arg: value
+ * @file: pointer to monty file
+ * @content: line content
+ * @lifi: flag change stack <-> queue
+ * Description: carries values through the program
+ */
 typedef struct bus_s
 {
 	char *arg;
@@ -58,9 +38,22 @@ typedef struct bus_s
 	int lifi;
 }  bus_t;
 extern bus_t bus;
-
+/**
+ * struct instruction_s - opcode and its function
+ * @opcode: the opcode
+ * @f: function to handle the opcode
+ *
+ * Description: opcode and its function
+ * for stack, queues, LIFO, FIFO Holberton project
+ */
+typedef struct instruction_s
+{
+	char *opcode;
+	void (*f)(stack_t **stack, unsigned int line_number);
+} instruction_t;
 char *_realloc(char *ptr, unsigned int old_size, unsigned int new_size);
 ssize_t getstdin(char **lineptr, int file);
+ssize_t getline(char **lineptr, size_t *n, FILE *stream);
 char  *clean_line(char *content);
 void f_push(stack_t **head, unsigned int number);
 void f_pall(stack_t **head, unsigned int number);
@@ -79,10 +72,8 @@ void f_pchar(stack_t **head, unsigned int counter);
 void f_pstr(stack_t **head, unsigned int counter);
 void f_rotl(stack_t **head, unsigned int counter);
 void f_rotr(stack_t **head, __attribute__((unused)) unsigned int counter);
-void _push(stack_t **head, int new);
+void addnode(stack_t **head, int n);
 void addqueue(stack_t **head, int n);
 void f_queue(stack_t **head, unsigned int counter);
 void f_stack(stack_t **head, unsigned int counter);
-
-
-#endif/*MONTY*/
+#endif
