+++ p_str.c
@@ -1,12 +1,10 @@
 #include "monty.h"
-
 /**
-* f_pstr - function that prints the string starting at the top of the stack,
-* followed by a new
-* @head: head of stack
-* @counter: line count
-*
-* Return: nothing
+ * f_pstr - prints the string starting at the top of the stack,
+ * followed by a new
+ * @head: stack head
+ * @counter: line_number
+ * Return: no return
 */
 void f_pstr(stack_t **head, unsigned int counter)
 {
