+++ pchar.c
@@ -1,12 +1,10 @@
 #include "monty.h"
-
 /**
-* f_pchar - function that prints the char at the top of the stack,
-* followed by a new line
-* @head: stack head
-* @counter: line count
-*
-* Return: nothing
+ * f_pchar - prints the char at the top of the stack,
+ * followed by a new line
+ * @head: stack head
+ * @counter: line_number
+ * Return: no return
 */
 void f_pchar(stack_t **head, unsigned int counter)
 {
