+++ queue.c
@@ -1,11 +1,9 @@
 #include "monty.h"
-
 /**
-* f_queue - function that prints the top of the queue
-* @head: head of queue
-* @counter: line count
-*
-* Return: nothing
+ * f_queue - prints the top
+ * @head: stack head
+ * @counter: line_number
+ * Return: no return
 */
 void f_queue(stack_t **head, unsigned int counter)
 {
@@ -15,17 +13,16 @@ void f_queue(stack_t **head, unsigned int counter)
 }
 
 /**
-* addqueue - function that add node to the tail the queue
-* @n: new value
-* @head: head of the queue
-*
-* Return: nothing
+ * addqueue - add node to the tail stack
+ * @n: new_value
+ * @head: head of the stack
+ * Return: no return
 */
 void addqueue(stack_t **head, int n)
 {
-	stack_t *new_node, *temp;
+	stack_t *new_node, *aux;
 
-	temp = *head;
+	aux = *head;
 	new_node = malloc(sizeof(stack_t));
 	if (new_node == NULL)
 	{
@@ -33,19 +30,19 @@ void addqueue(stack_t **head, int n)
 	}
 	new_node->n = n;
 	new_node->next = NULL;
-	if (temp)
+	if (aux)
 	{
-		while (temp->next)
-			temp = temp->next;
+		while (aux->next)
+			aux = aux->next;
 	}
-	if (!temp)
+	if (!aux)
 	{
 		*head = new_node;
 		new_node->prev = NULL;
 	}
 	else
 	{
-		temp->next = new_node;
-		new_node->prev = temp;
+		aux->next = new_node;
+		new_node->prev = aux;
 	}
 }
