+++ rotl.c
@@ -1,12 +1,10 @@
 #include "monty.h"
-
 /**
-* f_rotl- function that rotates the stack to the top
-* @head: head of the stack
-* @counter: line count
-*
-* Return: nothing
-*/
+  *f_rotl- rotates the stack to the top
+  *@head: stack head
+  *@counter: line_number
+  *Return: no return
+ */
 void f_rotl(stack_t **head,  __attribute__((unused)) unsigned int counter)
 {
 	stack_t *tmp = *head, *aux;
